requirements.txt
# Core Data Libraries
pandas==2.3.3
numpy==1.26.4

# Korean Stock Market
pykrx==1.2.3
finance-datareader==0.9.102

# Web Scraping / HTTP
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2

# Chart Generation
mplfinance==0.12.10b0
matplotlib==3.8.4

# Google Gemini AI
google-generativeai==0.8.6

# Utilities
openpyxl
python-dotenv==1.0.1
plyer==2.1.0
pocketbase==0.15.0
